speech-api/requirements.txt
flask
flask-cors
pydub
webvtt-py
nltk
yt-dlp
ffmpeg
gevent
tensorboardX
webrtcvad
fastapi[all]
